fix: normalize function description
